Asteroid Movement/Level Up:
Diagonal movement and randomized speed of asteroids added.  Initialization of level up complete.  Each time the set limit of asteroids per level is cleared from the screen, the number of asteroids, the speed at which they move, and the speed at which they spawn all increase.  Still need to implement the graphic for when the level is cleared.